Added dao classes, db connection
Added dao interfaces and classes, new configuration for database connection. DataSource supplier, wrote read and update methods for Post objects. Added postPermalink field for comment to be able to join the tables in the db. Added postgres dependency, fixed minor bugs in the field naming.  Added sql scripts for the Comments and Posts tables for postgres.