Added initial versions of servlets and instruction page
The app will have all the logic(counting words, tickers and working on data) moved to another app. Servlet classes have been created, but are still non functional. instruction.html contains set of instructions that will be available for the client. The separation of logic will make the app inner working easier to maintain and understand.  the classes related to logic will still remain in the app until the other one is created.