Prepared app for servlet container deployment
The app is deployable to Tomcat servlet container. Added web.xml for more clear config, changed path retrieval for AppProperties to make it work in container. Added caching to RService and fixed HelpServlet and PostServlet. DataListener class has been added to run a thread fetching data from reddit and updating cache from database in the background. Still needs optimizations, tests will be done when the app is in it's somehow final form.